+++ example/github.ts
@@ -1,8 +1,9 @@
 import { openai } from '@ai-sdk/openai'
 import { tool } from 'ai'
 import { agent, execute } from 'flows-ai'
-import { forEach, oneOf, parallel, sequence } from 'flows-ai/flows'
 import z from 'zod'
+
+import { githubProjectHealthAnalysisFlow } from './flows'
 
 const communicationAgent = agent({
   model: openai('gpt-4o'),
@@ -68,56 +69,17 @@ const userInputAgent = agent({
   },
 })
 
-export const githubProjectHealthAnalysisFlow = sequence([
-  {
-    agent: 'userInputAgent',
-    name: 'getProjectName',
-    input: 'Get a valid Github project name in format "organization/project"',
+const response = await execute(githubProjectHealthAnalysisFlow, {
+  agents: {
+    userInputAgent,
+    githubAgent,
+    communicationAgent,
   },
-  {
-    agent: 'githubAgent',
-    name: 'getIssues',
-    input: 'Go to Github and get the top 3 most popular issues and number of open issues.',
+  onFlowStart: (flow) => {
+    if (flow.name) {
+      console.log('Executing', flow.name)
+    }
   },
-  forEach({
-    item: 'Github issue and total number of open issues',
-    input: parallel([
-      oneOf([
-        {
-          when: 'There are more than 500 open issues',
-          input: {
-            agent: 'communicationAgent',
-            input: 'Write an email to the maintainer saying he is behind schedule.',
-          },
-        },
-        {
-          when: 'There are less than 500 open issues',
-          input: {
-            agent: 'communicationAgent',
-            input: 'Inform the maintainer that he is doing good job.',
-          },
-        },
-      ]),
-      {
-        agent: 'communicationAgent',
-        name: 'informMaintainer',
-        input: 'Inform the maintainer about open issue.',
-      },
-    ]),
-  }),
-])
+})
 
-// const response = await execute(githubProjectHealthAnalysisFlow, {
-//   agents: {
-//     userInputAgent,
-//     githubAgent,
-//     communicationAgent,
-//   },
-//   onFlowStart: (flow) => {
-//     if (flow.name) {
-//       console.log('Executing', flow.name)
-//     }
-//   },
-// })
-
-// console.log('Received response', response)
+console.log('Received response', response)
